Move get_repo_loader function from cli to repository package
