+++ qwikstart/cli/resolver.py
@@ -12,7 +12,7 @@ def resolve_task(
     execution_config: Optional[Dict[str, Any]] = None,
 ) -> Task:
     try:
-        loader = get_repo_loader(task_path, repo_url)
+        loader = repository.get_repo_loader(task_path, repo_url)
     except RepoLoaderError as error:
         raise UserFacingError(str(error)) from error
 
@@ -22,11 +22,3 @@ def resolve_task(
     execution_config = execution_config or {}
     execution_config.setdefault("source_dir", loader.repo_path)
     return parse_task(task_spec, execution_config=execution_config)
-
-
-def get_repo_loader(
-    task_path: str, repo_url: Optional[str] = None
-) -> repository.BaseRepoLoader:
-    if repo_url is not None:
-        return repository.GitRepoLoader(repo_url, task_path)
-    return repository.RepoLoader(task_path)
