+++ qwikstart/repository/__init__.py
@@ -1,5 +1,5 @@
 from .core import OperationsList, OperationSpec, TaskSpec
-from .loaders import BaseRepoLoader, GitRepoLoader, RepoLoader
+from .loaders import BaseRepoLoader, GitRepoLoader, RepoLoader, get_repo_loader
 
 __all__ = [
     "BaseRepoLoader",
@@ -8,4 +8,5 @@ __all__ = [
     "OperationsList",
     "RepoLoader",
     "TaskSpec",
+    "get_repo_loader",
 ]
